arduino/tests/cicd_test/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
pyserial==3.5
requests==2.31.0
urllib3==2.2.1
